+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/levelgen/biome/BWGWorldGenConfig.java
@@ -26,7 +26,7 @@ import java.nio.file.Path;
 import java.util.Map;
 import java.util.function.Supplier;
 
-public record BWGWorldGenConfig(Map<ResourceKey<Biome>, Boolean> enabledBiomes, int regionWeight) {
+public record BWGWorldGenConfig(Map<ResourceKey<Biome>, Boolean> enabledBiomes, int regionWeight, boolean vanillaAdditions) {
 
     public static final Path PATH = PlatformHandler.PLATFORM_HANDLER.configPath().resolve("world_generation.json5");
 
@@ -36,7 +36,8 @@ public record BWGWorldGenConfig(Map<ResourceKey<Biome>, Boolean> enabledBiomes, 
     public static final Codec<BWGWorldGenConfig> CODEC = RecordCodecBuilder.create(instance ->
             instance.group(
                     CommentedCodec.of(Codec.unboundedMap(ResourceKey.codec(Registries.BIOME), Codec.BOOL), "enabled_biomes", "Which biomes are enabled, if disabled the biome will default to its vanilla counterpart for the given region").forGetter(BWGWorldGenConfig::enabledBiomes),
-                    CommentedCodec.of(Codec.INT, "region_weight", "How much each BWG region weighs. This weight applies to all 3 BWG Regions").forGetter(BWGWorldGenConfig::regionWeight)
+                    CommentedCodec.of(Codec.INT, "region_weight", "How much each BWG region weighs. This weight applies to all 3 BWG Regions").forGetter(BWGWorldGenConfig::regionWeight),
+                    CommentedCodec.of(Codec.BOOL, "vanilla_additons", "Whether to add bwg flowers and features to Vanilla Biomes (Config Option for Fabric Only)").forGetter(config -> true)
             ).apply(instance, BWGWorldGenConfig::new)
     );
 
@@ -48,7 +49,7 @@ public record BWGWorldGenConfig(Map<ResourceKey<Biome>, Boolean> enabledBiomes, 
 
         enabledBiomes.put(BWGBiomes.ERODED_BOREALIS, false);
 
-        return new BWGWorldGenConfig(enabledBiomes, 8);
+        return new BWGWorldGenConfig(enabledBiomes, 8, true);
     }
 
     public static BWGWorldGenConfig getOrCreateConfigFromDisk() {
@@ -81,7 +82,7 @@ public record BWGWorldGenConfig(Map<ResourceKey<Biome>, Boolean> enabledBiomes, 
                     temporary.putAll(defaultWorldGenConfig.enabledBiomes);
                     temporary.putAll(config.enabledBiomes);
 
-                    BWGWorldGenConfig toCreate = new BWGWorldGenConfig(temporary, config.regionWeight);
+                    BWGWorldGenConfig toCreate = new BWGWorldGenConfig(temporary, config.regionWeight, config.vanillaAdditions);
 
                     createDefaultFile(toCreate);